[stable] Update tests to use Xcode 16.2 and iOS 18.2 Simulator (#166789)
Impacted Users: Flutter Release team, no external users impacted

Impact Description: Runs tests on Xcode 16 and iOS 18 Simulator

Workaround: N/A

Risk: Low

Test Coverage: Yes

Validation Steps: N/A

+++ engine/src/flutter/testing/ios_scenario_app/bin/run_ios_tests.dart
@@ -64,6 +64,7 @@ void main(List<String> args) async {
         deviceIdentifier: results.option('device-identifier')!,
         osRuntime: results.option('os-runtime')!,
         osVersion: results.option('os-version')!,
+        withImpeller: results.flag('with-impeller'),
         dumpXcresultOnFailure: dumpXcresultOnFailurePath,
       );
       completer.complete();
@@ -110,6 +111,7 @@ Future<void> _run(
   required String deviceIdentifier,
   required String osRuntime,
   required String osVersion,
+  required bool withImpeller,
   required String? dumpXcresultOnFailure,
 }) async {
   // Terminate early on SIGINT.
@@ -130,35 +132,37 @@ Future<void> _run(
 
   cleanup.add(() => _deleteIfPresent(resultBundle));
 
-  final process = await _runTests(
-    outScenariosPath: scenarioPath,
-    resultBundlePath: resultBundle.path,
-    osVersion: osVersion,
-    deviceName: deviceName,
-    iosEngineVariant: iosEngineVariant,
-  );
-  cleanup.add(process.kill);
-
-  // Create a temporary directory, if needed.
-  var storePath = dumpXcresultOnFailure;
-  if (storePath == null) {
-    final dumpDir = io.Directory.systemTemp.createTempSync();
-    storePath = dumpDir.path;
-    cleanup.add(() => dumpDir.delete(recursive: true));
-  }
-
-  if (await process.exitCode != 0) {
-    final String outputPath = _zipAndStoreFailedTestResults(
+  if (withImpeller) {
+    final process = await _runTests(
+      outScenariosPath: scenarioPath,
+      resultBundlePath: resultBundle.path,
+      osVersion: osVersion,
+      deviceName: deviceName,
       iosEngineVariant: iosEngineVariant,
-      resultBundle: resultBundle,
-      storePath: storePath,
     );
-    io.stderr.writeln('Failed test results are stored at $outputPath');
-    throw _ToolFailure('test failed.');
-  } else {
-    io.stderr.writeln('test succcess.');
+    cleanup.add(process.kill);
+
+    // Create a temporary directory, if needed.
+    var storePath = dumpXcresultOnFailure;
+    if (storePath == null) {
+      final dumpDir = io.Directory.systemTemp.createTempSync();
+      storePath = dumpDir.path;
+      cleanup.add(() => dumpDir.delete(recursive: true));
+    }
+
+    if (await process.exitCode != 0) {
+      final String outputPath = _zipAndStoreFailedTestResults(
+        iosEngineVariant: iosEngineVariant,
+        resultBundle: resultBundle,
+        storePath: storePath,
+      );
+      io.stderr.writeln('Failed test results are stored at $outputPath');
+      throw _ToolFailure('test failed.');
+    } else {
+      io.stderr.writeln('test succcess.');
+    }
+    _deleteIfPresent(resultBundle);
   }
-  _deleteIfPresent(resultBundle);
 }
 
 /// Exception thrown when the tool should halt execution intentionally.
@@ -193,6 +197,11 @@ final _args =
         'os-version',
         help: 'The OS version of the iOS simulator device to use.',
         defaultsTo: '18.2',
+      )
+      ..addFlag(
+        'with-impeller',
+        help: 'Whether to use the Impeller backend to run the tests.',
+        defaultsTo: true,
       )
       ..addOption(
         'dump-xcresult-on-failure',
@@ -256,7 +265,7 @@ void _createDevice({
   required String iosEngineVariant,
 }) {
   final scenarioPath = path.normalize(
-    path.join(engine.outDir.path, iosEngineVariant, 'ios_scenario_app', 'Scenarios'),
+    path.join(engine.outDir.path, iosEngineVariant, 'scenario_app', 'Scenarios'),
   );
 
   // Create a temporary directory to store the test results.
